Fixed some styles in [TextTheme]
* moved text based themes to separated file
