+++ lib/src/theme.dart
@@ -1,5 +1,8 @@
+import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 import 'package:yaru/src/utils/colors.dart' as yaru;
+
+part 'text_theme.dart';
 
 final _lightColorScheme = ColorScheme.fromSwatch(
     // NOTE(robert-ancell): Light shades from 'Tint' on website, dark shades calculated.
@@ -42,42 +45,6 @@ final _darkColorScheme = ColorScheme.fromSwatch(
     backgroundColor: yaru.Colors.coolGrey,
     errorColor: yaru.Colors.red,
     brightness: Brightness.dark);
-
-const _textTheme = TextTheme(
-  headline1: _UbuntuTextStyle(
-    fontSize: 96,
-    letterSpacing: -1.5,
-    fontWeight: FontWeight.w300,
-  ),
-  headline2: _UbuntuTextStyle(
-    fontSize: 60,
-    letterSpacing: -0.5,
-    fontWeight: FontWeight.w300,
-  ),
-  headline3: _UbuntuTextStyle(fontSize: 48),
-  headline4: _UbuntuTextStyle(fontSize: 34),
-  headline5: _UbuntuTextStyle(fontSize: 24, letterSpacing: -0.18),
-  headline6: _UbuntuTextStyle(
-    fontSize: 20,
-    letterSpacing: 0.15,
-    fontWeight: FontWeight.w500,
-  ),
-  subtitle1: _UbuntuTextStyle(fontSize: 16, letterSpacing: 0.15),
-  subtitle2: _UbuntuTextStyle(
-    fontSize: 14,
-    letterSpacing: 0.1,
-    fontWeight: FontWeight.w500,
-  ),
-  bodyText1: _UbuntuTextStyle(fontSize: 16, letterSpacing: 0.5),
-  bodyText2: _UbuntuTextStyle(fontSize: 14, letterSpacing: 0.25),
-  button: _UbuntuTextStyle(
-    fontSize: 14,
-    letterSpacing: 0.25,
-    fontWeight: FontWeight.w500,
-  ),
-  caption: _UbuntuTextStyle(fontSize: 12, letterSpacing: 0.4),
-  overline: _UbuntuTextStyle(fontSize: 10, letterSpacing: 1.5),
-);
 
 final _buttonThemeData = ButtonThemeData(
     shape: RoundedRectangleBorder(borderRadius: BorderRadius.circular(4.0)));
@@ -286,22 +253,3 @@ final _checkStyle = CheckboxThemeData(
 // Radios
 final _radioStyle = RadioThemeData(
     fillColor: MaterialStateProperty.resolveWith(_getCheckFillColorDark));
-
-class _UbuntuTextStyle extends TextStyle {
-  const _UbuntuTextStyle({
-    double? fontSize,
-    double? letterSpacing,
-    FontWeight? fontWeight,
-  }) : super(
-          fontSize: fontSize,
-          fontWeight: fontWeight,
-          letterSpacing: letterSpacing,
-          fontFamily: 'Ubuntu',
-          package: 'yaru',
-          fontFamilyFallback: const [
-            'Liberation Serif',
-            'Noto Sans CJK',
-            'Tibetan Machine Uni',
-          ],
-        );
-}
